Pin monolog/monolog to 2.2.0
Same version as in the master branch for WMF production.

This dependency became part of the tarball through the addition of
wikimedia/shellbox, however the version was not pinned, thus it ended
up changing as part of an unrelated commit I was drafting.

In previous release branches, this would be pruned after branching
from master, but this time around it stayed for wikimedia/shellbox,
but wasn't pinned.

Change-Id: I51995058c0e1807c84b88d79e82e4a4005d63a13